refactor: 정적파일 핸들링 방식 변경
