+++ tomcat/src/main/java/org/apache/coyote/http11/controller/StaticResourceController.java
@@ -1,17 +1,11 @@
 package org.apache.coyote.http11.controller;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import org.apache.coyote.http11.http.ContentType;
+import org.apache.coyote.http11.controller.util.StaticResourceReader;
 import org.apache.coyote.http11.http.HttpStatus;
-import org.apache.coyote.http11.request.dto.HttpRequest;
-import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.http.request.dto.HttpRequest;
+import org.apache.coyote.http11.http.response.HttpResponse;
 
 public class StaticResourceController extends AbstractController {
-
-    private static final String DEFAULT_MIMETYPE = "application/octet-stream";
 
     private final String base;
     private final String defaultDocument;
@@ -26,29 +20,12 @@ public class StaticResourceController extends AbstractController {
         String resourcePath = request.path().replaceFirst("^/static/?", "");
         String normalizedResourcePath = normalizePath(resourcePath);
         String fullPath = base + "/" + normalizedResourcePath;
+        byte[] bytes = StaticResourceReader.readResource(fullPath);
+        String contentType = StaticResourceReader.resolveContentType(normalizedResourcePath);
 
-        try (InputStream resourceStream = getClass().getClassLoader().getResourceAsStream(fullPath)) {
-            if (resourceStream == null) {
-                // FIXME: 예외처리해서 404 처리 하는게 좋것 같아
-                response.status(HttpStatus.NOT_FOUND)
-                        .contentType(ContentType.PLAIN.value())
-                        .write("404 Not Found");
-                return;
-            }
-
-            byte[] bytes = resourceStream.readAllBytes();
-            String contentType = resolveContentType(normalizedResourcePath);
-
-            response.status(HttpStatus.OK)
-                    .contentType(contentType)
-                    .write(bytes);
-
-        } catch (IOException e) {
-            // FIXME: 예외처리해서 404 처리 하는게 좋것 같아
-            response.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .contentType(ContentType.PLAIN.value())
-                    .write("500 Internal Server Error");
-        }
+        response.status(HttpStatus.OK)
+                .contentType(contentType)
+                .write(bytes);
     }
 
     private String normalizePath(String path) {
@@ -58,19 +35,6 @@ public class StaticResourceController extends AbstractController {
         if (path.isEmpty()) {
             path = defaultDocument;
         }
-        if (!path.contains(".")) {
-            path = path + ".html";
-        }
         return path;
-    }
-
-    private String resolveContentType(String resourcePath) {
-        try {
-            Path path = Path.of(resourcePath);
-            String mimeType = Files.probeContentType(path);
-            return (mimeType != null ? mimeType : DEFAULT_MIMETYPE);
-        } catch (IOException e) {
-            return DEFAULT_MIMETYPE;
-        }
     }
 }
