+++ tomcat/src/main/java/org/apache/coyote/http11/handler/statics/StaticResourceHandlerMapping.java
@@ -1,8 +1,9 @@
-package org.apache.coyote.http11.handler.ready;
+package org.apache.coyote.http11.handler.statics;
 
 import org.apache.coyote.http11.controller.Controller;
 import org.apache.coyote.http11.controller.StaticResourceController;
-import org.apache.coyote.http11.request.dto.HttpRequest;
+import org.apache.coyote.http11.handler.HandlerMapping;
+import org.apache.coyote.http11.http.request.dto.HttpRequest;
 
 public class StaticResourceHandlerMapping implements HandlerMapping {
 
@@ -13,6 +14,9 @@ public class StaticResourceHandlerMapping implements HandlerMapping {
 
     @Override
     public Controller getHandler(HttpRequest request) {
+        if (!request.hasGet()) {
+            return null;
+        }
         if (hasStaticResource(request.path())) {
             return new StaticResourceController("static", "index.html");
         }
@@ -21,8 +25,18 @@ public class StaticResourceHandlerMapping implements HandlerMapping {
 
     private boolean hasStaticResource(String path) {
         String resourcePath = path.replaceFirst("^/static/?", "");
-        if (resourcePath.isEmpty()) resourcePath = "index.html";
-        String fullPath = "static/" + resourcePath;
+        String normalizedResourcePath = normalizePath(resourcePath);
+        String fullPath = "static/" + normalizedResourcePath;
         return getClass().getClassLoader().getResource(fullPath) != null;
+    }
+
+    private String normalizePath(String path) {
+        if (path.startsWith("/")) {
+            path = path.substring(1);
+        }
+        if (path.isEmpty()) {
+            path = "index.html";
+        }
+        return path;
     }
 }
